[FEATURE-5] Move CurrentUser to Entities folder
